changed dio data type to allow for device update flag bit

+++ inc/Devices.h
@@ -9,7 +9,7 @@
 #include <math.h>
 
 typedef void (*init_t)(ini_t*, unsigned int);
-typedef void (*update_t)(unsigned short*, int*);
+typedef void (*update_t)(unsigned int*, int*);
 typedef void (*unload_t)();
 
 init_t *inits;
@@ -24,7 +24,7 @@ unsigned int unloadCount = 0;
 unsigned int nextDIO = 0; //perhaps replace this small system and the current HINSTANCE handling with a struct containing HINSTANCE and its assigned dio port?
 
 //DIO functions
-void initDevices(unsigned short *dio, unsigned int dioCount, ini_t *cfg) {
+void initDevices(unsigned int *dio, unsigned int dioCount, ini_t *cfg) {
 	devices = calloc(0, sizeof(HINSTANCE));
 	inits = calloc(1, sizeof(init_t));
 	updates = calloc(1, sizeof(update_t));
@@ -60,7 +60,7 @@ void initDevices(unsigned short *dio, unsigned int dioCount, ini_t *cfg) {
 	}
 }
 
-void updateDevices(unsigned short *dio, unsigned short *flgReg, unsigned short *intReg, int *execute) {
+void updateDevices(unsigned int *dio, unsigned short *flgReg, unsigned short *intReg, int *execute) {
 	//implement a means to signal a device that its dio register has been modified by the CPU
 
 	for(unsigned int i = 0; i < updateCount && updates[i]; i++) updates[i](dio, execute);
